svn path=/branches/v0.6.1-NewMeta/; revision=1279

+++ Source/ZEEngine/ZEGame/ZEScene.cpp
@@ -51,10 +51,10 @@
 #include "ZEPhysics/ZEPhysicalWorld.h"
 #include "ZEGraphics/ZEFrameRenderer.h"
 #include "ZEGraphics/ZEShadowRenderer.h"
-#include "ZEMap/ZEPortalMap/ZEPortalMap.h"
+#include "ZEInterior/ZEInterior.h"
 #include "ZESerialization/ZEFileSerializer.h"
 #include "ZESerialization/ZEFileUnserializer.h"
-#include "ZEMap/ZEPortalMap/ZEPortalMapResource.h"
+#include "ZEInterior/ZEInteriorResource.h"
 
 #include <memory.h>
 
@@ -85,7 +85,8 @@ static ZEString ConstructResourcePath(const ZEString& Path)
 
 bool ZEScene::Initialize()
 {
-	Deinitialize();
+	if (Initialized)
+		return true;
 
 	if (Renderer == NULL)
 		Renderer = ZEFrameRenderer::CreateInstance();
@@ -105,7 +106,7 @@ bool ZEScene::Initialize()
 	if (ShadowRenderer == NULL)
 		ShadowRenderer = ZEShadowRenderer::CreateInstance();
 
-	if (Renderer == NULL)
+	if (ShadowRenderer == NULL)
 	{
 		zeCriticalError("Can not create shadow renderer.");
 		return false;
@@ -132,9 +133,11 @@ bool ZEScene::Initialize()
 		return false;
 	}
 
-
 	for (ZESize I = 0; I < Entities.GetCount(); I++)
+	{
+		Entities[I]->SetOwnerScene(this);
 		Entities[I]->Initialize();
+	}
 
 	Initialized = true;
 	return true;
@@ -142,6 +145,9 @@ bool ZEScene::Initialize()
 
 void ZEScene::Deinitialize()
 {
+	if (!Initialized)
+		return;
+
 	for (ZESize I = 0; I < Entities.GetCount(); I++)
 		Entities[I]->Deinitialize();
 
@@ -187,36 +193,31 @@ void ZEScene::Destroy()
 
 void ZEScene::AddEntity(ZEEntity* Entity)
 {
+	if (Entities.Exists(Entity))
+	{
+		zeError("Can not add an already native entity.");
+		return;
+	}
+
 	Entity->SetEntityId(LastEntityId++);
 	Entities.Add(Entity);
 	Entity->SetOwnerScene(this);
 
-	ZEArray<ZEEntity*> Temp = Entity->GetComponents();
-
-	for (ZESize I = 0; I < Temp.GetCount(); I++)
-		Temp[I]->SetOwnerScene(this);
-
-	Temp = Entity->GetChildEntities();
-
-	for (ZESize I = 0; I < Temp.GetCount(); I++)
-		Temp[I]->SetOwnerScene(this);
+	if(Entity->GetName().GetLength() == 0)
+		Entity->SetName(Entity->GetClass()->GetName() + ZEString(LastEntityId));
 
 	Entity->Initialize();
 }
 
 void ZEScene::RemoveEntity(ZEEntity* Entity)
 {
+	if (!Entities.Exists(Entity))
+	{
+		zeError("Can not remove a foreign entity.");
+		return;
+	}
+
 	Entity->SetOwnerScene(NULL);
-
-	ZEArray<ZEEntity*> Temp = Entity->GetComponents();
-
-	for (ZESize I = 0; I < Temp.GetCount(); I++)
-		Temp[I]->SetOwnerScene(NULL);
-
-	Temp = Entity->GetChildEntities();
-
-	for (ZESize I = 0; I < Temp.GetCount(); I++)
-		Temp[I]->SetOwnerScene(NULL);
 
 	Entities.DeleteValue(Entity);
 }
@@ -255,6 +256,11 @@ ZERenderer* ZEScene::GetRenderer()
 	return Renderer;
 }
 
+ZERenderer*	ZEScene::GetShadowRenderer()
+{
+	return ShadowRenderer;
+}
+
 ZEPhysicalWorld* ZEScene::GetPhysicalWorld()
 {
 	return PhysicalWorld;
@@ -276,7 +282,12 @@ void ZEScene::SetActiveListener(ZEListener* Listener)
 	zeSound->SetActiveListener(Listener);
 }
 
-const ZECullStatistics& ZEScene::GetCullerStatistics()
+ZESceneCuller& ZEScene::GetSceneCuller()
+{
+	return Culler;
+}
+
+const ZESceneStatistics& ZEScene::GetStatistics() const
 {
 	return Culler.GetStatistics();
 }
@@ -295,11 +306,13 @@ void ZEScene::Tick(ZEEntity* Entity, float ElapsedTime)
 	
 	const ZEArray<ZEEntity*>& Components = Entity->GetComponents();
 	for (ZESize N = 0; N < Components.GetCount(); N++)
-		Components[N]->Tick(ElapsedTime);
+		Tick(Components[N], ElapsedTime);
+
 
 	const ZEArray<ZEEntity*>& SubEntities = Entity->GetChildEntities();
 	for (ZESize N = 0; N < SubEntities.GetCount(); N++)
-		SubEntities[N]->Tick(ElapsedTime);
+		Tick(SubEntities[N], ElapsedTime);
+
 }
 
 void ZEScene::Tick(float ElapsedTime)
@@ -315,16 +328,21 @@ void ZEScene::Render(float ElapsedTime)
 
 	Renderer->SetCamera(ActiveCamera);
 	
-	ZEDrawParameters DrawParameters;
-	DrawParameters.ElapsedTime = ElapsedTime;
-	DrawParameters.FrameId = zeCore->GetFrameId();
-	DrawParameters.Pass = ZE_RP_COLOR;
-	DrawParameters.Renderer = Renderer;
-	DrawParameters.ViewVolume = (ZEViewVolume*)&ActiveCamera->GetViewVolume();
-	DrawParameters.View = (ZEView*)&ActiveCamera->GetView();
-	DrawParameters.Lights.Clear();
+	FrameDrawParameters.ElapsedTime = ElapsedTime;
+	FrameDrawParameters.FrameId = zeCore->GetFrameId();
+	FrameDrawParameters.Pass = ZE_RP_COLOR;
+	FrameDrawParameters.Renderer = Renderer;
+	FrameDrawParameters.ViewVolume = (ZEViewVolume*)&ActiveCamera->GetViewVolume();
+	FrameDrawParameters.View = (ZEView*)&ActiveCamera->GetView();
+	FrameDrawParameters.Lights.Clear();
 
-	Culler.CullScene(this, &DrawParameters);
+	memset(&FrameDrawParameters.Statistics, 0, sizeof(ZEDrawStatistics));
+
+	FrameDrawParameters.Renderer->SetDrawParameters(&FrameDrawParameters);
+
+	Culler.CullScene(this, &FrameDrawParameters);
+
+	//Fill Draw Parameters Statistics ZESceneStats section
 }
 
 bool ZEScene::Save(const ZEString& FileName)
@@ -433,6 +451,26 @@ bool ZEScene::Load(const ZEString& FileName)
 	}
 }
 
+void ZEScene::SetAmbientFactor(float Factor)
+{
+	AmbientFactor = Factor;
+}
+
+float ZEScene::GetAmbientFactor() const
+{
+	return AmbientFactor;
+}
+
+void ZEScene::SetAmbientColor(ZEVector3 Color)
+{
+	AmbientColor = Color;
+}
+
+const ZEVector3& ZEScene::GetAmbientColor() const
+{
+	return AmbientColor;
+}
+
 ZEScene::ZEScene()
 {
 	Initialized = false;
@@ -443,11 +481,16 @@ ZEScene::ZEScene()
 	ActiveCamera = NULL;
 	ActiveListener = NULL;
 	PhysicalWorld = NULL;
+	AmbientColor = ZEVector3::One;
+	AmbientFactor = 0.0f;
 }
 
 ZEScene::~ZEScene()
 {
 	Deinitialize();
+
+	for (ZESize I = 0; I < Entities.GetCount(); I++)
+		Entities[I]->Destroy();
 }
 
 ZEScene* ZEScene::GetInstance()
